feat: allow forcing rust libm
