+++ packages/vexide-core/Cargo.toml
@@ -34,6 +34,7 @@ bitflags = "2.4.2"
 futures-core = { version = "0.3.30", default-features = false, features = ["alloc"] }
 pin-project = "1.1.5"
 replace_with = { version = "0.1.7", default-features = false }
+libm = { version = "0.2.8", optional = true }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 dlmalloc = { version = "0.2.4", features = ["global"] }
@@ -41,3 +42,6 @@ libm = "0.2.8"
 
 [lints]
 workspace = true
+
+[features]
+force_rust_libm = ["dep:libm"]
